+++ modules/sde-submodules/part-site-information-as-planned/src/main/java/org/eclipse/tractusx/sde/submodels/psiap/steps/EDCPartSiteInformationAsPlannedHandlerStep.java
@@ -24,15 +24,12 @@ import java.util.Map;
 import org.eclipse.tractusx.sde.common.constants.CommonConstants;
 import org.eclipse.tractusx.sde.common.entities.PolicyModel;
 import org.eclipse.tractusx.sde.common.exception.CsvHandlerUseCaseException;
-import org.eclipse.tractusx.sde.common.exception.ServiceException;
 import org.eclipse.tractusx.sde.common.submodel.executor.Step;
 import org.eclipse.tractusx.sde.edc.entities.request.asset.AssetEntryRequest;
 import org.eclipse.tractusx.sde.edc.entities.request.asset.AssetEntryRequestFactory;
 import org.eclipse.tractusx.sde.edc.facilitator.CreateEDCAssetFacilator;
 import org.eclipse.tractusx.sde.edc.gateways.external.EDCGateway;
-import org.eclipse.tractusx.sde.submodels.psiap.entity.PartSiteInformationAsPlannedEntity;
 import org.eclipse.tractusx.sde.submodels.psiap.model.PartSiteInformationAsPlanned;
-import org.eclipse.tractusx.sde.submodels.psiap.services.PartSiteInformationAsPlannedService;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
@@ -45,7 +42,6 @@ public class EDCPartSiteInformationAsPlannedHandlerStep extends Step {
 	private final AssetEntryRequestFactory assetFactory;
 	private final EDCGateway edcGateway;
 	private final CreateEDCAssetFacilator createEDCAssetFacilator;
-	private final PartSiteInformationAsPlannedService partSiteInformationAsPlannedService;
 
 	@SneakyThrows
 	public PartSiteInformationAsPlanned run(String submodel, PartSiteInformationAsPlanned input, String processId,
@@ -57,46 +53,25 @@ public class EDCPartSiteInformationAsPlannedHandlerStep extends Step {
 
 			AssetEntryRequest assetEntryRequest = assetFactory.getAssetRequest(submodel,
 					getSubmodelShortDescriptionOfModel(), shellId, subModelId, input.getUuid());
-			if (!edcGateway.assetExistsLookup(assetEntryRequest.getId())) {
-				edcProcessingforPartAsPlanned(assetEntryRequest, input, policy);
-			} else {
 
-				deleteEDCFirstForUpdate(submodel, input, processId);
-				edcProcessingforPartAsPlanned(assetEntryRequest, input, policy);
+			Map<String, String> eDCAsset = null;
+
+			if (!edcGateway.assetExistsLookup(assetEntryRequest.getId())) {
+				eDCAsset = createEDCAssetFacilator.createEDCAsset(assetEntryRequest, policy);
+			} else {
+				eDCAsset = createEDCAssetFacilator.updateEDCAsset(assetEntryRequest, policy);
 				input.setUpdated(CommonConstants.UPDATED_Y);
 			}
+
+			// EDC transaction information for DB
+			input.setAssetId(eDCAsset.get("assetId"));
+			input.setAccessPolicyId(eDCAsset.get("accessPolicyId"));
+			input.setUsagePolicyId(eDCAsset.get("usagePolicyId"));
+			input.setContractDefinationId(eDCAsset.get("contractDefinitionId"));
 
 			return input;
 		} catch (Exception e) {
 			throw new CsvHandlerUseCaseException(input.getRowNumber(), "EDC: " + e.getMessage());
 		}
 	}
-
-	@SneakyThrows
-	private void deleteEDCFirstForUpdate(String submodel, PartSiteInformationAsPlanned input, String processId) {
-		try {
-			PartSiteInformationAsPlannedEntity partSiteInformationAsPlannedEntity = partSiteInformationAsPlannedService
-					.readEntity(input.getUuid());
-			partSiteInformationAsPlannedService.deleteEDCAsset(partSiteInformationAsPlannedEntity);
-
-		} catch (Exception e) {
-			if (!e.getMessage().contains("404 Not Found")) {
-				throw new ServiceException("Unable to delete EDC offer for update: " + e.getMessage());
-			}
-		}
-	}
-
-	@SneakyThrows
-	private void edcProcessingforPartAsPlanned(AssetEntryRequest assetEntryRequest, PartSiteInformationAsPlanned input,
-			PolicyModel policy) {
-
-		Map<String, String> createEDCAsset = createEDCAssetFacilator.createEDCAsset(assetEntryRequest, policy);
-
-		// EDC transaction information for DB
-		input.setAssetId(assetEntryRequest.getId());
-		input.setAccessPolicyId(createEDCAsset.get("accessPolicyId"));
-		input.setUsagePolicyId(createEDCAsset.get("usagePolicyId"));
-		input.setContractDefinationId(createEDCAsset.get("contractDefinitionId"));
-	}
-
 }