+++ modules/sde-submodules/pcf/src/main/java/org/eclipse/tractusx/sde/submodels/pcf/entity/PcfEntity.java
@@ -126,6 +126,7 @@ public class PcfEntity extends CommonPropEntity {
 	@Column(name = "extTFS_distributionStageDlucGhgEmissions")
 	private Double extTFSDistributionStageDlucGhgEmissions; // 1.0,
 	private Double distributionStagePcfIncludingBiogenic; // 0.0,
+	private Double distributionStageAircraftGhgEmissions; // 0,
 	private Double carbonContentBiogenic; // 0.0,
 	private String partialFullPcf;// Cradle-to-gate
 	private String productId;// urn:gtin:4712345060507